src/requirements.txt
openai==0.27.10
Flask==2.3.2
PyMuPDF==1.22.3
langchain==0.0.314
docarray==0.32.1
pandoc==2.3
pypandoc==1.11
wikipedia==1.4.0
google-search-results==2.4.2
google-api-python-client==2.88.0
streamlit
streamlit-chat==0.1.1
streamlit-extras==0.2.7
unstructured==0.8.1
redis==4.5.5
faiss-cpu==1.7.4
tiktoken==0.3.3
rank_bm25==0.2.2
html2text==2020.1.16
python-pptx==0.6.21
doctran
tenacity==8.2.3
sounddevice==0.4.6
soundfile==0.12.1
# elevenlabs==0.2.26
pip-tools==7.3.0
streamlit-modal==0.1.0
st-pages==0.4.4
docxtpl==0.16.7
langchain_experimental==0.0.29
playwright==1.38.0
SQLAlchemy==2.0.22
spacy==3.5.4
tokenizers==0.13.3
urllib3==2.0.4
requests==2.31.0
rank-bm25==0.2.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
python-xlib==0.33
pydantic==1.10.9
# feast==0.31.1
beautifulsoup4==4.12.2
loguru==0.7.0
pynput==1.7.6
PyYAML
st-click-detector
st_clickable_images
tenacity==8.2.3
toml==0.10.2
geopy==2.4.0
httpx-oauth
streamlit-authenticator
extra-streamlit-components
localstoragePy
hydralit-components
streamlit-datalist
opensearch-py
requests-aws4auth
lancedb
pywinctl
streamlit-audiorec
streamlit-mic-recorder
streamlit-bokeh-events
apify-client
streamlit_js_eval